cloud_functions/setup-gmail-watch/requirements.txt
functions-framework==3.*
google-api-python-client
google-cloud-storage
google-cloud-secret-manager
requests